requirements.txt
joblib>=1.3.1
numba>=0.53.1
pandas>=2.0.0
pretty_midi>=0.2.10
tokenizers>=0.13.3
torch>=2.0.0
transformers>=4.29.2
lightning>=2.0.0
beautifulsoup4

muster @ git+https://github.com/TimFelixBeyer/amtevaluation.github.io
music21 @ git+https://github.com/TimFelixBeyer/music21@0ed70bb
score_transformer @ git+https://github.com/TimFelixBeyer/ScoreTransformer@934a228
